+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 
+## Unreleased yet
+
+ADDED:
+
+- Allow middlewares to execute on "not found" routes.
+
 ## 1.4.8
 
 ADDED:
+
 - Middlewares are now also executed for OPTIONS requests.
 - Update OPTIONS HTTP Request logic so that the CORS headers are only included
   in the response when the "--cors" flag is used. If "--cors" is not used,